refactor(tui): split tui.rs into a module by responsibility
Replace the single tui.rs file with a tui/ module:
- state.rs   — AppState, RunState, TranscriptRole, TranscriptEntry
- composer.rs — new_composer, composer_height
- terminal.rs — AppTerminal, init_terminal, restore_terminal
- app.rs      — TuiApp struct and event loop
- mod.rs      — public execute entry point

Tests stay co-located with the code they cover.
