+++ apps/cli/src/commands/tui/terminal.rs
@@ -1,0 +1,41 @@
+use std::io::{self, Stdout};
+
+use anyhow::{Context, Result};
+use crossterm::{
+    execute,
+    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
+};
+use ratatui::{backend::CrosstermBackend, Terminal};
+
+pub(crate) type AppTerminal = Terminal<CrosstermBackend<Stdout>>;
+
+pub(crate) fn init_terminal() -> Result<AppTerminal> {
+    let original_hook = std::panic::take_hook();
+    std::panic::set_hook(Box::new(move |info| {
+        let _ = disable_raw_mode();
+        let _ = execute!(io::stdout(), LeaveAlternateScreen);
+        original_hook(info);
+    }));
+
+    enable_raw_mode().context("failed to enable raw mode")?;
+    let mut stdout = io::stdout();
+
+    if let Err(e) = execute!(stdout, EnterAlternateScreen) {
+        let _ = disable_raw_mode();
+        return Err(e).context("failed to enter alternate screen");
+    }
+
+    let backend = CrosstermBackend::new(stdout);
+    Terminal::new(backend).map_err(|e| {
+        let _ = disable_raw_mode();
+        let _ = execute!(io::stdout(), LeaveAlternateScreen);
+        anyhow::anyhow!("failed to initialize terminal: {e}")
+    })
+}
+
+pub(crate) fn restore_terminal(terminal: &mut AppTerminal) -> Result<()> {
+    disable_raw_mode().context("failed to disable raw mode")?;
+    execute!(terminal.backend_mut(), LeaveAlternateScreen)
+        .context("failed to leave alternate screen")?;
+    terminal.show_cursor().context("failed to restore cursor")
+}
